Define replacement maps for a keyword.
A keyword can now have a replacement map, which maps a string to a numeric
value. The interpreter will then understand the string and convert it back
to the number.

This is usefull for having boolean parameters in the rendering settings
(where "on"/"off" can be mapped to 1/0), or enumerated parameters (where
options map to indices).

As an example, the channel color can now be given by a triple (r, g, b),
or by a string, e.g. "red", which maps to the triple (255, 0, 0).