Rename AutoCompletion to AnimationAutoCompletion.
AutoCompletion previously re-implemented the original AutoCompletion class
from the AutoComplete library.

Since Fiji first loads the original AutoCompletion class, the
re-implemented version is not used (and it also shouldn't since it might
affect Fiji's script editor).

So instead, we have the same class renamed and extending AutoCompletion.
This is hacky, but it seems to work.